@svick 2018-03-18
@dotnet-bot test Linux x64 Release Build please
@dotnet-bot test OSX x64 Debug Build please