@karelz 2018-03-18
Do we still need the original issue? Is it still tracking a meaningful improvement over your latest change?

cc @valenis @brianrob